project/apriltag/table_guidance_project_v8/requirements.txt
numpy>=1.23
opencv-python>=4.7
pyyaml>=6.0
pupil-apriltags>=1.0.4
ultralytics>=8.0.0
